test/requirements.txt
torch
transformers
pandas
openpyxl
scikit-learn
google-generativeai
